Add files via upload
 3 - 5 Hours

We interact with banking systems all the time without thinking about it: when we pay for our groceries with a debit card, when our employer sends us a paycheck, or when we apply for a loan. Banks are perhaps the most foundational service for our global, connected world - and as such, they have incredibly complex systems. In this mini-project, you'll engineer a simple model version of a banking system with the Python OOP techniques you've been learning. You'll have a great degree of freedom in working on this project, it will be up to you to define data entities and choose the format you'll export your data into.